src/requirements.txt
psycopg~=3.2.3
pytxc~=0.1.6
python-dotenv~=1.0.1
bods_client~=0.13.0
